Set default local path value instead of as hint
Using a hint is problematic in that the user can think that it will
be used as the default value when it will not and it also prevents
the normal field validation from presenting the user with an error.
Making the value for the local path derivded from the remote as the
default pre-filled into the edittext seems to make more sense and
more closely matches the git cli behaviour.

Fixes: #209